requirements-minimal.txt
-e git+git@github.com:jackfriedson/bitex.git@master#egg=BitEx
click>=6.7,<7.0
coinbase>=2.0.6,<3.0.0
ipdb>=0.10.3,<1.0.0
pandas>=0.21.0,<1.0.0
pyYAML>=3.12.0,<4.0.0
requests>=2.18.4,<3.0.0
tzlocal>=1.5.1,<2.0.0
